Simplified yr_test
